DNN-5943 - refactored workflow setting. Totally moved to the tab package
